% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_matrix_H.R
\name{auxiliary_matrix_H}
\alias{auxiliary_matrix_H}
\title{Create an Auxiliary Matrix H}
\usage{
auxiliary_matrix_H(c, n)
}
\arguments{
\item{c}{A numeric value that will fill the diagonals just off the main diagonal.}

\item{n}{An integer specifying the size of the square matrix \(H\) (the number of rows and columns).}
}
\value{
A numeric matrix \(H\) of size \(n * n\), where the diagonal elements are 1, and the elements just off the main diagonal are set to the value \(c\).
}
\description{
This function creates an auxiliary matrix \(H\) of size \(n * n\) with ones on the diagonal, and a constant value \(c\) on the diagonals just off the main diagonal (except for the first and last row/column).
}
\examples{
\dontrun{
# Example: Create an auxiliary matrix H with n = 5 and c = 0.5
n <- 5
c <- 0.5
H <- auxiliary_matrix_H(c, n)
print(H)
}
}
