% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{f3}
\alias{f3}
\title{Generate Random Numbers from Uniform Distribution}
\usage{
f3(num)
}
\arguments{
\item{num}{An integer specifying the number of random numbers to generate}
}
\value{
A NumericMatrix of size \code{num x 1}, where each element is a random number between 0 and 100
}
\description{
This function generates a matrix of random numbers drawn from a uniform distribution between 0 and 100
}
\examples{
\dontrun{
# Generate 5 random numbers between 0 and 100
random_numbers <- f3(5)
print(random_numbers)
}
}
