% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_of_XTHXQ.R
\name{rank_of_XTHXQ}
\alias{rank_of_XTHXQ}
\title{Compute the Rank of \( X^T H X Q \)}
\usage{
rank_of_XTHXQ(X, H, Q)
}
\arguments{
\item{X}{A numeric matrix of size \(n * m\), where \(n\) is the number of rows and \(m\) is the number of columns.}

\item{H}{A numeric matrix of size \(n * n\), which is an auxiliary matrix used in the calculation.}

\item{Q}{A numeric matrix of size \(m * m\), which is another auxiliary matrix used in the calculation.}
}
\value{
An integer representing the rank of the matrix \( X^T H X Q \).
}
\description{
This function computes the rank of the matrix \( X^T H X Q \), where \( X \) is a matrix, \( H \) is an auxiliary matrix, and \( Q \) is another auxiliary matrix. The rank is computed using the \code{qr()} function, which performs a QR decomposition.
}
\examples{
\dontrun{
# Example: Compute the rank of X^T H X Q
X <- matrix(rnorm(100), nrow = 10)
H <- auxiliary_matrix_H(0.5, 10)
Q <- auxiliary_matrix_Q(10)
rank_result <- rank_of_XTHXQ(X, H, Q)
print(rank_result)
}
}
